Feature/business logic

+++ src/main/java/com/stockmanagement/application/services/SupplierService.java
@@ -8,7 +8,7 @@ public interface SupplierService {
 
     SupplierDto create(SupplierDto supplierDTO);
     List<SupplierDto> get();
-    SupplierDto getById(Long id);
-    SupplierDto update(Long id, SupplierDto updated);
-    void delete(Long id);
+    SupplierDto getById(Integer id);
+    SupplierDto update(Integer id, SupplierDto updated);
+    void delete(Integer id);
 }
